backend/requirements.txt
flask==1.1.4
pyjwt==1.7.1
PyMySQL==1.0.2
scrypt==0.8.18
flask_sqlalchemy==2.5.1
flask_script==2.0.6
flask_migrate==2.5.3
python-dotenv==0.19.0
pyyaml==5.4.1
gunicorn==20.1.0
requests==2.22.0
markdown==3.3.6
markupsafe==2.0.1